+++ tests/Integration/Service.Tests/PartsModuleSpecs/ContextBase.cs
@@ -3,11 +3,10 @@
     using System.Collections.Generic;
     using System.Security.Claims;
 
-    using Domain.LinnApps.Parts;
-
     using Linn.Common.Facade;
     using Linn.Common.Persistence;
     using Linn.Stores.Domain.LinnApps;
+    using Linn.Stores.Domain.LinnApps.Parts;
     using Linn.Stores.Facade.ResourceBuilders;
     using Linn.Stores.Resources;
     using Linn.Stores.Service.Modules;
@@ -21,12 +20,12 @@
 
     public class ContextBase : NancyContextBase
     {
-        protected IFacadeService<Part, long, PartResource, PartResource> PartsFacadeService
+        protected IFacadeService<Part, int, PartResource, PartResource> PartsFacadeService
         {
             get; private set;
         }
 
-        protected IRepository<Part, long> PartRepository { get; private set; }
+        protected IRepository<Part, int> PartRepository { get; private set; }
 
         protected IRepository<ParetoClass, string> ParetoClassRepository { get; private set; }
 
@@ -37,9 +36,9 @@
         public void EstablishContext()
         {
             this.PartsFacadeService = Substitute
-                .For<IFacadeService<Part, long, PartResource, PartResource>>();
+                .For<IFacadeService<Part, int, PartResource, PartResource>>();
 
-            this.PartRepository = Substitute.For<IRepository<Part, long>>();
+            this.PartRepository = Substitute.For<IRepository<Part, int>>();
             this.ParetoClassRepository = Substitute.For<IRepository<ParetoClass, string>>();
             this.ProductAnalysisCodeRepository = Substitute.For<IRepository<ProductAnalysisCode, string>>();
